requirements.txt
python-dotenv
aiogram
mysql-connector-python
